+++ changelog.md
@@ -1,2 +1,1 @@
-- Fixed corpse instantly despawning when taking out items
-- Fixed corpse gui still showing up even if the corpse despawned
+- Fixed history gui closing bug